+++ ihaskell-hvega/CHANGELOG.md
@@ -4,8 +4,9 @@ For the latest version of this document, please see
 ## 0.2.0.0
 
 Added the `IHaskell.Display.HvegaLab` module for use with Jupyter Lab,
-which has native VegaLite support. Try it out in [Tweag's jupyterWith
-environment](https://github.com/tweag/jupyterWith).
+which has native VegaLite support. It requires the use of the `vlShow`
+helper (needed to avoid an overlapping instance from `IHaskell.Display.Hveva`).
+Try it out in [Tweag's jupyterWith environment](https://github.com/tweag/jupyterWith).
 
 ## 0.1.0.3