+++ ihaskell-hvega/CHANGELOG.md
@@ -1,6 +1,13 @@
 For the latest version of this document, please see
 [https://github.com/DougBurke/hvega/blob/master/ihaskell-hvega/CHANGELOG.md](https://github.com/DougBurke/hvega/blob/master/ihaskell-hvega/CHANGELOG.md).
 
+## 0.2.0.0
+
+Added the `vlShow` helper to allow Vega-Lite visualizations to be
+viewed directly in Jupyter lab (rather than Jupyter notebook).
+
+Try it out in [Tweag's jupyterWith environment](https://github.com/tweag/jupyterWith).
+
 ## 0.1.0.3
 
 The only change is to the cabal file, where `cabal-version: >=1.10` has